some small fixes

+++ draw.cpp
@@ -40,7 +40,7 @@ void draw_rectangle(std::vector<uint32_t> &image, const uint32_t &color, const u
     }
 }
 
-void print_progress(uint8_t frame, uint8_t frames) {
+void print_progress(uint16_t frame, uint16_t frames) {
     int adjusted_frame = static_cast<int>(((float)frame / frames) * 100);
     std::string progress = "[" + std::string(adjusted_frame, '=') + std::string(100 - adjusted_frame, ' ') + "]";
     std::cout << "\r\033[F" << adjusted_frame << "%\n" << progress << std::flush;
@@ -48,7 +48,7 @@ void print_progress(uint8_t frame, uint8_t frames) {
 }
 
 void animation(SDL_Renderer* SDLrenderer, Render* renderer, Movement m, uint32_t frames) {
-    for (uint8_t frame = 1; frame <= frames; frame++) {
+    for (uint16_t frame = 1; frame <= frames; frame++) {
         renderer->move(m);
         renderer->main_render(SDLrenderer);
         std::stringstream ss;
